backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
openai>=1.0.0
anthropic>=0.7.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=2.0.0
jinja2>=3.0.0
stripe>=5.0.0
paypalrestsdk>=1.7.0
pymysql>=1.0.0
mysql-connector-python>=8.0.0
cryptography>=41.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
httpx[cli]>=0.24.0
aiofiles>=23.0.0
python-json-logger>=2.0.0
prometheus-client>=0.17.0
